Move Staleness map to be internal
So users are forced to use the correct methods.

Also adds lots of documentation
