+++ internal/exp/metrics/staleness/staleness_test.go
@@ -9,47 +9,6 @@ import (
 	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/exp/metrics/identity"
 )
 
-// RawMap
-var _ Map[time.Time] = (*RawMap[identity.Stream, time.Time])(nil)
-
-type RawMap[K comparable, V any] map[K]V
-
-func (rm *RawMap[K, V]) Load(key K) (V, bool) {
-	value, ok := (*rm)[key]
-	return value, ok
-}
-
-func (rm *RawMap[K, V]) LoadOrStore(key K, value V) (V, bool) {
-	returnedVal, ok := (*rm)[key]
-	if !ok {
-		(*rm)[key] = value
-		returnedVal = value
-	}
-
-	return returnedVal, ok
-}
-
-func (rm *RawMap[K, V]) Store(key K, value V) {
-	(*rm)[key] = value
-}
-
-func (rm *RawMap[K, V]) Delete(key K) {
-	delete(*rm, key)
-}
-
-func (rm *RawMap[K, V]) Items() func(yield func(K, V) bool) bool {
-	return func(yield func(K, V) bool) bool {
-		for k, v := range *rm {
-			if !yield(k, v) {
-				break
-			}
-		}
-		return false
-	}
-}
-
-// Tests
-
 func TestStaleness(t *testing.T) {
 	t.Parallel()
